+++ nutree/typed_tree.py
@@ -14,6 +14,7 @@ Declare the :class:`~nutree.tree.TypedTree` class.
 
 from __future__ import annotations
 
+import copy
 import warnings
 from collections import Counter
 from collections.abc import Iterator
@@ -26,11 +27,13 @@ from typing_extensions import Any, Self
 from nutree.common import (
     ROOT_DATA_ID,
     ROOT_NODE_ID,
+    CalcIdCallbackType,
+    CycleDetectedError,
     DataIdType,
     DeserializeMapperType,
+    DotMapperCallbackType,
     IterMethod,
     KeyMapType,
-    MapperCallbackType,
     PredicateCallbackType,
     SerializeMapperType,
     UniqueConstraintError,
@@ -84,11 +87,12 @@ class TypedNode(Node[TData]):
         node_id: int | None = None,
         meta: dict | None = None,
     ):
-        self._kind: str = kind  # tree._register() checks for this attribute
+        # tree._register() checks for this attribute in __init__():
+        self._kind: str = kind
         super().__init__(
             data, parent=parent, data_id=data_id, node_id=node_id, meta=meta
         )
-        assert isinstance(kind, str) and kind != ANY_KIND, f"Unsupported `kind`: {kind}"
+        assert isinstance(kind, str), f"Unsupported `kind`: {kind}"
 
         # del self._children
         # self._child_map: Dict[Node] = None
@@ -265,7 +269,9 @@ class TypedNode(Node[TData]):
 
         assert not self._children
         for child in other.children:
-            new_child = self.add_child(child.data, kind=None, data_id=child._data_id)
+            new_child = self.add_child(
+                child.data, kind=child.kind, data_id=child._data_id
+            )
             if child.children:
                 new_child._add_from(child, predicate=None)
         return
@@ -492,7 +498,7 @@ class TypedNode(Node[TData]):
     def copy(
         self, *, add_self: bool = True, predicate: PredicateCallbackType | None = None
     ) -> TypedTree[TData]:
-        """Return a new :class:`~nutree.tree.Tree` instance from this branch.
+        """Return a new :class:`~nutree.typed_tree.TypedTree` instance from this branch.
 
         See also :ref:`iteration-callbacks`.
         """
@@ -531,8 +537,8 @@ class TypedNode(Node[TData]):
         graph_attrs: dict | None = None,
         node_attrs: dict | None = None,
         edge_attrs: dict | None = None,
-        node_mapper: MapperCallbackType | None = None,
-        edge_mapper: MapperCallbackType | None = None,
+        node_mapper: DotMapperCallbackType | None = None,
+        edge_mapper: DotMapperCallbackType | None = None,
     ) -> Iterator[str]:
         """Generate a DOT formatted graph representation.
 
@@ -597,6 +603,21 @@ class TypedTree(Tree[TData, TypedNode[TData]]):
     #: Default value for ``add_child`` when loading.
     DEFAULT_CHILD_TYPE = "child"
 
+    def __init__(
+        self,
+        name: str | None = None,
+        *,
+        calc_data_id: CalcIdCallbackType | None = None,
+        forward_attrs: bool = False,
+    ) -> None:
+        super().__init__(
+            name=name,
+            calc_data_id=calc_data_id,
+            forward_attrs=forward_attrs,
+            check_dag=True,
+        )
+        self._system_root = self.root_node_factory(self)
+
     @classmethod
     def deserialize_mapper(cls, parent: Node, data: dict) -> str | object | None:
         """Used as default `mapper` argument for :meth:`load`."""
@@ -607,6 +628,62 @@ class TypedTree(Tree[TData, TypedNode[TData]]):
         raise NotImplementedError(
             f"Override this method or pass a mapper callback to evaluate {data}."
         )
+
+    def deepcopy(
+        self,
+        *,
+        name: str | None = None,
+        memo: dict[int, Any] | None = None,
+    ) -> TypedTree[TData]:
+        """Return a deep copy of this tree.
+
+        New :class:`~nutree.typed_tree.TypedTree`
+        and :class:`~nutree.typed_tree.TypedNode` instances are created.
+        The new nodes reference deep-copied data objects.
+
+        See Node's :meth:`~nutree.node.Node.copy_to` and :ref:`iteration-callbacks`
+        method for details.
+        """
+        if name is None:
+            name = f"Deep copy of {self}"
+        if memo is None:
+            memo = {}
+        new_tree = self.__class__(name)
+
+        def _copy_children(source_parent, target_parent) -> None:
+            for c in source_parent.children:
+                new_data = copy.deepcopy(c.data, memo)
+                new_data_id = new_tree.calc_data_id(new_data)
+                new_node = target_parent.add(
+                    new_data, deep=False, data_id=new_data_id, kind=c.kind
+                )
+                _copy_children(c, new_node)
+
+        with self:
+            _copy_children(self.system_root, new_tree.system_root)
+
+        return new_tree
+
+    def _check_insert(self, node: Node):
+        """Raise error if inserting a node would violate DAG restrictions."""
+        # We can assume that node.parent is set and that node already has at
+        # least one clone registered in self._nodes_by_data_id, when this is
+        # called from _register()
+        assert node._kind, node
+        ref_key = node._data_id
+        kind = node._kind
+        if node._parent._children:
+            for sibling in node._parent._children:
+                if sibling._data_id == ref_key and sibling._kind == kind:
+                    raise UniqueConstraintError(
+                        f"Node with data_id {ref_key} and kind {kind} "
+                        f"already exists in parent {node._parent}"
+                    )
+        for n in self._nodes_by_data_id[ref_key]:
+            if node.is_descendant_of(n) and n._kind == kind:
+                raise CycleDetectedError(
+                    f"Inserting {node} would create a cycle with {n}"
+                )
 
     def add_child(
         self,