+++ CHANGELOG.md
@@ -1,19 +1,28 @@
 # Changelog
 
+**NOTE** 
+Dropping Python versions that reached EOL will only cause a minor version bump.
+
 ## 1.2.0 (unreleased)
+- BREAKING: Drop Python 3.9 (EOL 2025-10)
 - BREAKING:
   - Some methods now raise `TypeError` instead of `ValueError` when 
     parameters have the wrong type or are missing.
-    (4fd93e98b99f8c6aa66d05d3b2e9fb22b5c185f2)
+- BREAKING: Rename `MapperCallbackType` to `DotMapperCallbackType`
 - DEPRECATE: `Tree.find()` is deprecated, use `Tree.find_first()` instead."
-- DEPRECATE: Calling method `{name}()` via `__getattr__` is deprecated.
+- DEPRECATE: Calling method `{name}()` by adding `()` to the attribute name via `__getattr__` is deprecated.
   This allowed calling simple methods from within `repr=TEMPLATE` parameter
    e.g. `repr="{node.path()}"`. Use `repr=CALLBACK` instead.
 - `tree[x]` now also accepts `Node` instances.
 - `x in tree` now uses the same lookup logic as `tree[x]`.
 - `reversed(tree)` now raises a TypError, since it didn't make sense.
-- New method `tree.sorted()` returns a sorted copy.
-- Drop Python 3.9 (EOL 2025-10)
+- Add method `tree.sorted()` returns a sorted copy.
+- Add `tree.deepcopy()`, so now we support `copy.copy(tree)` and `copy.deepcopy(tree)`.
+- `len(node)` returns number of direct children. 
+  However nodes are always truthy, even if they don't have children.
+- Add method `node.parent_iterator()`.
+- Add option `Tree(..., check_dag=True)` to enforce/relax DAG compliance (default: true).
+- Add method `tree.map(fn)`
 - Support Python 3.14, 3.15
 - Use [uv](https://docs.astral.sh/uv/)